# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HustleCoin V2 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mining/status:
    get:
      operationId: getMiningStatus
      tags: [mining]
      summary: Get current mining session status
      responses:
        "200":
          description: Mining status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiningStatus"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mining/start:
    post:
      operationId: startMining
      tags: [mining]
      summary: Start a new 24-hour mining session
      responses:
        "200":
          description: Mining session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiningStatus"
        "400":
          description: Session already active or not claimable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mining/claim:
    post:
      operationId: claimMining
      tags: [mining]
      summary: Claim rewards from a completed mining session
      responses:
        "200":
          description: Rewards claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiningClaimResult"
        "400":
          description: Session not complete or no active session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mining/history:
    get:
      operationId: getMiningHistory
      tags: [mining]
      summary: Get paginated mining history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Mining history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiningHistoryResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    MiningStatus:
      type: object
      required: [state, balance, streak, totalMines, miningRate, estimatedReward]
      properties:
        state:
          type: string
          enum: [idle, mining, claimable]
          description: "idle = no session; mining = session active, timer running; claimable = session complete, ready to claim"
        sessionStartedAt:
          type: ["string", "null"]
          description: ISO timestamp when session started
        sessionEndsAt:
          type: ["string", "null"]
          description: ISO timestamp when session completes (sessionStartedAt + 24h)
        secondsRemaining:
          type: ["integer", "null"]
          description: Seconds left in the current session
        balance:
          type: integer
        streak:
          type: integer
        totalMines:
          type: integer
        miningRate:
          type: number
          description: HP earned per hour in current/next session
        estimatedReward:
          type: integer
          description: Total HP that will be earned when session completes
        lastClaimedAt:
          type: ["string", "null"]
    MiningClaimResult:
      type: object
      required: [hpEarned, bonusHp, totalReward, newBalance, newStreak, newTotalMines]
      properties:
        hpEarned:
          type: integer
          description: Base HP earned
        bonusHp:
          type: integer
          description: Streak bonus HP
        totalReward:
          type: integer
          description: Total HP added to balance
        newBalance:
          type: integer
        newStreak:
          type: integer
        newTotalMines:
          type: integer
    MiningLogEntry:
      type: object
      required: [id, hpEarned, bonusHp, totalHp, streak, minedAt]
      properties:
        id:
          type: integer
        hpEarned:
          type: integer
        bonusHp:
          type: integer
        totalHp:
          type: integer
        streak:
          type: integer
        minedAt:
          type: string
    MiningHistoryResponse:
      type: object
      required: [entries, total]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/MiningLogEntry"
        total:
          type: integer
...